+++ src/components/PresenceMap/component.js
@@ -1,9 +1,10 @@
 import React, { useRef, useState, useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types';
-import getMapContainerStyles from './getMapContainerStyles';
-import StyledContainer from './styled/container';
 import { usePrevious } from '../../libs/hooks';
+import createMarkers from './helpers/createMarkers';
+import StyledContainer from './styled/container';
 import StyledControls from './styled/controls';
+import getMapContainerStyles from './getMapContainerStyles';
 
 const MAX_ZOOM = 15;
 
@@ -24,90 +25,18 @@ export default function MapComponent(props) {
     const [map, setMap] = useState(null);
 
     const initMap = useCallback(() => {
-        const getImage = async url =>
-            new Promise((resolve, reject) => {
-                const img = new Image();
-                img.crossOrigin = 'anonymous';
-                img.onload = () => resolve(img);
-                img.onerror = reject;
-                img.src = url;
-            });
-
-        const getBase64Image = async (srcUrl, heading) => {
-            const img = await getImage(srcUrl);
-            const canvas = document.createElement('canvas');
-            const ctx = canvas.getContext('2d');
-            canvas.width = 40;
-            canvas.height = 50;
-            const cache = img;
-            const imageWidth = cache.width;
-            const imageHeight = cache.height;
-            ctx.save();
-            ctx.clearRect(0, 0, canvas.width, canvas.height);
-            ctx.translate(canvas.width / 2, canvas.height / 2);
-            ctx.rotate((Math.PI / 180) * heading);
-            ctx.translate(-(canvas.width / 2), -(canvas.height / 2));
-            const xCenter = canvas.width / 2 - imageWidth / 2;
-            const yCenter = canvas.height / 2 - imageHeight / 2;
-            ctx.drawImage(img, xCenter, yCenter, imageWidth, imageHeight);
-            ctx.restore();
-
-            return canvas.toDataURL();
-        };
-
         const mapOptions = {
             zoom: zoom === 'auto' ? MAX_ZOOM : zoom,
             center: center !== 'auto' ? center : null,
             disableDefaultUI: true,
         };
-        const newMap = new window.google.maps.Map(mapContainer.current, mapOptions);
-        const markersCopy = [];
-
-        objects.forEach(marker => {
-            const heading = marker.heading ? marker.heading : 0;
-
-            const markerOptions = {
-                position: marker.position,
-                map: newMap,
-                zIndex: Math.round(marker.position.lat * 100000),
-            };
-
-            let promImage = null;
-            if (marker.image) {
-                promImage = new Promise((resolve, reject) => {
-                    const result = getBase64Image(marker.image, heading);
-                    result
-                        .then(img => {
-                            markerOptions.icon = img;
-                            resolve(true);
-                        })
-                        .catch(error => {
-                            reject(`Error creating new image. Details: ${error}`);
-                        });
-                });
-            } else {
-                promImage = new Promise(resolve => resolve(true));
-            }
-
-            promImage
-                .then(() => {
-                    const markerMap = new window.google.maps.Marker(markerOptions);
-                    markersCopy.push(markerMap);
-
-                    if (center === 'auto') {
-                        const bounds = new window.google.maps.LatLngBounds();
-                        markersCopy.forEach(markerObj => bounds.extend(markerObj.getPosition()));
-                        newMap.setCenter(bounds.getCenter());
-                        newMap.fitBounds(bounds);
-                    }
-                })
-                .catch(error => {
-                    // eslint-disable-next-line no-console
-                    console.log(`Error. Details: ${error}`);
-                });
+        const googleMap = new window.google.maps.Map(mapContainer.current, mapOptions);
+        createMarkers({
+            objects,
+            googleMap,
+            center,
         });
-
-        setMap(newMap);
+        setMap(googleMap);
     }, [objects, center, zoom]);
 
     const prevIsScriptLoaded = usePrevious(props.isScriptLoaded);
@@ -181,13 +110,13 @@ MapComponent.propTypes = {
     ),
     className: PropTypes.string,
     style: PropTypes.object,
-    zoom: PropTypes.oneOf([PropTypes.number, 'auto']),
-    center: PropTypes.oneOf([
+    zoom: PropTypes.oneOfType([PropTypes.number, PropTypes.oneOf(['auto'])]),
+    center: PropTypes.oneOfType([
         PropTypes.shape({
             lat: PropTypes.number,
             lng: PropTypes.number,
         }),
-        'auto',
+        PropTypes.oneOf(['auto']),
     ]),
     showTraffic: PropTypes.bool,
     showTransit: PropTypes.bool,
